requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.2
django-js-asset==2.2.0
django-ratelimit==4.1.0
django-tinymce==4.1.0
Faker==30.8.1
idna==3.10
mysqlclient==2.2.5
pillow==11.0.0
psycopg==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
